+++ src/app/shared/question-dropdown.ts
@@ -1,4 +1,4 @@
-import { QuestionBase } from './question.model';
+import { QuestionBase } from './question-base';
 
 export class DropdownQuestion extends QuestionBase<string> {
   controlType = 'dropdown';
