+++ src/app/shared/question-textbox.ts
@@ -1,4 +1,4 @@
-import { QuestionBase } from './question.model';
+import { QuestionBase } from './question-base';
 
 export class TextboxQuestion extends QuestionBase<string> {
   controlType = 'textbox';
